Refactored simple Book class to an Entity for coop with db
