+++ src/main/java/home/andreybelonog/controllers/BookController.java
@@ -1,6 +1,6 @@
 package home.andreybelonog.controllers;
 
-import home.andreybelonog.Book;
+import home.andreybelonog.domain.Book;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,7 +15,6 @@ public class BookController {
 
     @GetMapping("api/book")
     public Book getBook(@RequestParam(value = "name", defaultValue = "DefaultBook") String name){
-        return new Book(counter.incrementAndGet(),
-                        String.format(TEMPLATE, name));
+        return null;
     }
 }
